do these checks firstly when we start hyperd
